Add files via upload
Expense Tracker

This project is an Expense Tracker application built using Qt for the GUI and Firebase for user authentication and data storage. The application allows users to:

- Sign in using their email and password through Firebase Authentication.
- Add Expenses by specifying the category, amount, date, and details of each expense.
- View Expenses by fetching and displaying them from the Firebase Realtime Database.

The project demonstrates the integration of Qt with Firebase, utilizing network requests to interact with Firebase Authentication and the Realtime Database, enabling a seamless and secure user experience.